refactor(core): extract parsing to indexers, relationships to resolver
The Section actors had grown to own both raw Mach-O parsing and
RuntimeObject translation, and RuntimeEngine carried the cross-image
relationship union inline. Realign the layering so each type has one
job:

- RuntimeObjCInterfaceIndexer takes the MachOImage at init and owns
  all ObjC/ObjCDump extraction in prepare(); RuntimeObjCSection is
  left as a thin RuntimeObject translator.
- RuntimeSwiftInterfaceIndexer wraps the upstream SwiftInterfaceIndexer
  and absorbs the relationship reverse-table build that previously
  ran inline in RuntimeSwiftSection.init.
- RuntimeRelationshipsResolver owns the cross-image subclass/conformer
  union; RuntimeEngine.relationships(for:) keeps only the thin
  local/remote dispatch.

Indexer owns raw data and parsing, Section translates to domain
types, Resolver owns cross-image logic.
